+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/di/module/MainActivityModule.kt
@@ -10,12 +10,4 @@ class MainActivityModule {
 
     @Provides
     fun provideMainActivityView(mainActivity: MainActivity): MainActivityView = mainActivity
-
-//    @Provides
-//    fun provideMainViewModel(activity: AppCompatActivity): MainViewModel =
-//        ViewModelProvider(activity).get(MainViewModel::class.java)
-//
-//    @Provides
-//    fun provideCitiesViewModel(activity: AppCompatActivity): CitiesViewModel =
-//        ViewModelProvider(activity).get(CitiesViewModel::class.java)
 }
