+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/di/module/UtilModule.kt
@@ -1,7 +1,7 @@
 package by.itacademy.training.task9mvp.di.module
 
-import by.itacademy.training.task9mvp.util.CurrentTemperatureUnitListener
-import by.itacademy.training.task9mvp.util.CurrentTemperatureUnitListenerImpl
+import by.itacademy.training.task9mvp.ui.presenter.CurrentTemperatureUnitListener
+import by.itacademy.training.task9mvp.ui.presenter.CurrentTemperatureUnitListenerImpl
 import by.itacademy.training.task9mvp.util.SupportSharedPreference
 import by.itacademy.training.task9mvp.util.SupportSharedPreferenceImpl
 import dagger.Binds
@@ -12,7 +12,7 @@ interface UtilModule {
 
     @Binds
     fun bindCurrentTemperatureUnitListener(listener: CurrentTemperatureUnitListenerImpl):
-        CurrentTemperatureUnitListener
+            CurrentTemperatureUnitListener
 
     @Binds
     fun bindSupportSharePreference(support: SupportSharedPreferenceImpl):
