+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/model/api/WeatherApi.kt
@@ -2,13 +2,14 @@ package by.itacademy.training.task9mvp.model.api
 
 import by.itacademy.training.task9mvp.model.dto.api.ApiResponse
 import io.reactivex.Observable
+import io.reactivex.Single
 import retrofit2.http.GET
 import retrofit2.http.Query
 
 interface WeatherApi {
 
     @GET("forecast.json?key=$WEATHER_API_KEY")
-    fun getCurrentWeather(@Query("q") city: String): Observable<ApiResponse>
+    fun getCurrentWeather(@Query("q") city: String): Single<ApiResponse>
 
     companion object {
         const val BASE_URL = "http://api.weatherapi.com/v1/"
