+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/model/domain/City.kt
@@ -1,0 +1,3 @@
+package by.itacademy.training.task9mvp.model.domain
+
+data class City(val cityName: String)
