+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/model/domain/ForecastDay.kt
@@ -1,3 +1,3 @@
-package by.itacademy.training.task9mvp.model.entity
+package by.itacademy.training.task9mvp.model.domain
 
 data class ForecastDay(val list: List<HourTemperature>)
