+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/model/domain/HourTemperature.kt
@@ -1,6 +1,4 @@
-package by.itacademy.training.task9mvp.model.entity
-
-import by.itacademy.training.task9mvp.model.entity.Condition
+package by.itacademy.training.task9mvp.model.domain
 
 data class HourTemperature(
     val celsiusTemperature: Int,
