+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/model/domain/TemperatureUnit.kt
@@ -1,4 +1,4 @@
-package by.itacademy.training.task9mvp.util
+package by.itacademy.training.task9mvp.model.domain
 
 enum class TemperatureUnit {
     FAHRENHEIT,
