+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/model/domain/WeatherReport.kt
@@ -1,4 +1,4 @@
-package by.itacademy.training.task9mvp.model.entity
+package by.itacademy.training.task9mvp.model.domain
 
 data class WeatherReport(
     val currentTemperature: CurrentTemperature,
