--- task9mvp/src/main/java/by/itacademy/training/task9mvp/model/entity/City.kt
@@ -1,3 +1,0 @@
-package by.itacademy.training.task9mvp.model.entity
-
-data class City(val cityName: String)
