+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/model/repository/CitiesRepository.kt
@@ -1,6 +1,6 @@
 package by.itacademy.training.task9mvp.model.repository
 
-import by.itacademy.training.task9mvp.model.entity.City
+import by.itacademy.training.task9mvp.model.domain.City
 import io.reactivex.Completable
 import io.reactivex.Observable
