+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/model/repository/WeatherForecastRepository.kt
@@ -1,9 +1,9 @@
 package by.itacademy.training.task9mvp.model.repository
 
-import by.itacademy.training.task9mvp.model.entity.WeatherReport
-import io.reactivex.Observable
+import by.itacademy.training.task9mvp.model.domain.WeatherReport
+import io.reactivex.Single
 
 interface WeatherForecastRepository {
 
-    fun getWeatherForecastForDay(city: String): Observable<WeatherReport>
+    fun getWeatherForecastForDay(city: String): Single<WeatherReport>
 }
