+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/ui/presenter/CitiesActivityPresenterImpl.kt
@@ -1,6 +1,6 @@
 package by.itacademy.training.task9mvp.ui.presenter
 
-import by.itacademy.training.task9mvp.model.entity.City
+import by.itacademy.training.task9mvp.model.domain.City
 import by.itacademy.training.task9mvp.model.repository.CitiesRepository
 import by.itacademy.training.task9mvp.ui.view.CityActivityView
 import by.itacademy.training.task9mvp.util.SupportSharedPreference
