+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/ui/presenter/MainActivityPresenter.kt
@@ -1,7 +1,7 @@
 package by.itacademy.training.task9mvp.ui.presenter
 
-import by.itacademy.training.task9mvp.model.entity.CurrentTemperature
-import by.itacademy.training.task9mvp.model.entity.WeatherReport
+import by.itacademy.training.task9mvp.model.domain.CurrentTemperature
+import by.itacademy.training.task9mvp.model.domain.WeatherReport
 
 interface MainActivityPresenter {
