+++ task9mvp/src/main/java/by/itacademy/training/task9mvp/ui/view/CityActivityView.kt
@@ -1,6 +1,6 @@
 package by.itacademy.training.task9mvp.ui.view
 
-import by.itacademy.training.task9mvp.model.entity.City
+import by.itacademy.training.task9mvp.model.domain.City
 
 interface CityActivityView {
