--- task9mvp/src/main/java/by/itacademy/training/task9mvp/util/Event.kt
@@ -1,22 +1,0 @@
-package by.itacademy.training.task9mvp.util
-
-class Event <out T>(
-    val status: Status,
-    val data: T?,
-    val message: String?
-) {
-    companion object {
-
-        fun <T> success(data: T): Event<T> = Event(Status.SUCCESS, data, null)
-
-        fun <T> loading(data: T?): Event<T> = Event(Status.LOADING, null, null)
-
-        fun <T> error(data: T?, message: String?): Event<T> = Event(Status.ERROR, null, message)
-    }
-}
-
-enum class Status {
-    SUCCESS,
-    LOADING,
-    ERROR
-}
